+++ src/main/scala/com/ideal/linked/toposoid/knowledgebase/featurevector/model/SingleFeatureVectorForSearch.scala
@@ -26,7 +26,7 @@ import play.api.libs.json.Json
  * @param epsilon Epsilon is used to determines how much to expand from search candidate radius.
  * @param timeout Timeout is used for search time deadline. The unit is nano-seconds.
  */
-case class SingleFeatureVectorForSearch(vector:List[Float], num:Int, radius:Float, epsilon:Float, timeout:Int)
+case class SingleFeatureVectorForSearch(vector:List[Float], num:Int, radius:Float, epsilon:Float, timeout:Long)
 object SingleFeatureVectorForSearch {
   implicit val jsonWrites = Json.writes[SingleFeatureVectorForSearch]
   implicit val jsonReads = Json.reads[SingleFeatureVectorForSearch]
